+++ internal/dataplane/parser/translate_tlsroute.go
@@ -29,11 +29,6 @@ func (p *Parser) ingressRulesFromTLSRoutes() ingressRules {
 
 	var errs []error
 	for _, tlsroute := range tlsRouteList {
-		if p.featureFlags.ExpressionRoutes {
-			p.registerResourceFailureNotSupportedForExpressionRoutes(tlsroute)
-			continue
-		}
-
 		if err := p.ingressRulesFromTLSRoute(&result, tlsroute); err != nil {
 			err = fmt.Errorf("TLSRoute %s/%s can't be routed: %w", tlsroute.Namespace, tlsroute.Name, err)
 			errs = append(errs, err)
@@ -42,6 +37,10 @@ func (p *Parser) ingressRulesFromTLSRoutes() ingressRules {
 			// reported as successfully parsed.
 			p.registerSuccessfullyParsedObject(tlsroute)
 		}
+	}
+
+	if p.featureFlags.ExpressionRoutes {
+		applyExpressionToIngressRules(&result)
 	}
 
 	if len(errs) > 0 {